Cadastro de Plantas

+++ php/planta_list.php
@@ -14,7 +14,10 @@
             echo '<td>';
                 echo '<a href="alterar_planta.php?id_plan='.$result[$i]["id_planta"].'"><i class="far fa-edit" title="Alterar"></i></a>';
                 echo '<a href="detalhes_planta.php?id_plan='.$result[$i]["id_planta"].'"><i class="fas fa-search-plus" title="Ver detalhes"></i></a>';
-                echo '<a href="php/planta_delete.php?id_plan='.$result[$i]["id_planta"].'"><i class="far fa-trash-alt" title="Excluir"></i></a>';      
+
+                if ($_SESSION["tipo"] == "A"){
+                   echo '<a href="php/planta_delete.php?id_plan='.$result[$i]["id_planta"].'"><i class="far fa-trash-alt" title="Excluir"></i></a>';      
+                }
             echo '</td>';
         echo '</tr>';
     }
